Parse Addresses

Parse Addresses

Usage

gc_address(data, address, city = NULL, zip = NULL, state = NULL, county = NULL)

Arguments

data

A data frame

address

The column containing the address. Can be a one-line address or just the street address component (if city, state, etc. are provided). Tidy-evaluated within data.

city

A column containing city names. Tidy-evaluated within data.

zip

A column containing ZIP codes. Tidy-evaluated within data.

state

A column containing states. Tidy-evaluated within data.

county

A column containing counties. Matched to an internal list and passed on to the output. Tidy-evaluated within data.

Value

A data frame containing columns with the parsed address. These match the Census format for street addresses.

Examples

gc_address(NULL, "125 5th St S St. Petersburg FL 33701")
#> # A tibble: 1 × 11
#>   state_code count…¹ zip_c…²   num num_s…³ stree…⁴ stree…⁵ stree…⁶ stree…⁷ unit 
#>   <chr>      <chr>   <chr>   <int> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>
#> 1 12         NA      33701     125 NA      NA      5TH     ST      S       NA   
#> # … with 1 more variable: city <chr>, and abbreviated variable names
#> #   ¹​county_code, ²​zip_code, ³​num_suff, ⁴​street_dir_pre, ⁵​street_name,
#> #   ⁶​street_type, ⁷​street_dir_suff
gc_address(NULL, "1225 St Charles St St Louis MO")
#> # A tibble: 1 × 11
#>   state_code count…¹ zip_c…²   num num_s…³ stree…⁴ stree…⁵ stree…⁶ stree…⁷ unit 
#>   <chr>      <chr>   <chr>   <int> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>
#> 1 29         NA      NA       1225 NA      NA      ST CHA… ST      NA      NA   
#> # … with 1 more variable: city <chr>, and abbreviated variable names
#> #   ¹​county_code, ²​zip_code, ³​num_suff, ⁴​street_dir_pre, ⁵​street_name,
#> #   ⁶​street_type, ⁷​street_dir_suff

data(nc_addr)
gc_address(nc_addr, address = address, city = city, zip = zip, county = county, state = "NC")
#> # A tibble: 1,000 × 11
#>    state_c…¹ count…² zip_c…³   num num_s…⁴ stree…⁵ stree…⁶ stree…⁷ stree…⁸ unit 
#>    <chr>     <chr>   <chr>   <int> <chr>   <chr>   <chr>   <chr>   <chr>   <chr>
#>  1 37        055     27948     505 NA      NA      BURNS   DR      NA      NA   
#>  2 37        055     27954     144 NA      NA      PAYNE   RD      NA      NA   
#>  3 37        055     27948     116 NA      NA      MYBET   CT      NA      NA   
#>  4 37        055     27949    4516 NA      NA      MARITY… LN      NA      NA   
#>  5 37        055     27959    2236 NA      S       CROATAN HWY     NA      NA   
#>  6 37        055     NA         NA NA      NA      NA      NA      NA      NA   
#>  7 37        055     27954     129 NA      NA      FIRST … CT      NA      NA   
#>  8 37        055     NA         NA NA      NA      NA      NA      NA      NA   
#>  9 37        055     27949    3634 NA      N       VA DARE TRL     NA      NA   
#> 10 37        055     27948    1911 NA      NA      HARBOU… DR      NA      NA   
#> # … with 990 more rows, 1 more variable: city <chr>, and abbreviated variable
#> #   names ¹​state_code, ²​county_code, ³​zip_code, ⁴​num_suff, ⁵​street_dir_pre,
#> #   ⁶​street_name, ⁷​street_type, ⁸​street_dir_suff